python-services/requirements.txt
# Core dependencies
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
scipy==1.11.4

# Face recognition and 3D reconstruction
face-recognition==1.3.0
mediapipe==0.10.8
trimesh==4.0.5
scikit-image==0.22.0

# 3D processing and avatar generation
pygltflib==1.16.0
bpy==4.0.0  # Blender Python API
makehuman==1.2.0  # Alternative: use MakeHuman Python bindings

# TTS (Hebrew support)
TTS==0.22.0  # Coqui TTS
piper-tts==1.2.0  # Piper TTS fallback
torch==2.1.1
torchaudio==2.1.1

# Wav2Lip dependencies
torch==2.1.1
torchvision==0.16.1
librosa==0.10.1
soundfile==0.12.1
resampy==0.4.2

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Web framework for Python services
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
